# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RingiFlow API
  description: ワークフロー管理システム RingiFlow の BFF API
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
paths:
...
  /api/v1/roles:
    get:
      tags:
      - roles
      summary: GET /api/v1/roles
      description: |-
        テナント内のロール一覧をユーザー数付きで取得する。
        system_admin は除外される。
      operationId: list_roles
      responses:
        '200':
          description: ロール一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleItemData'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
    post:
      tags:
      - roles
      summary: POST /api/v1/roles
      description: カスタムロールを作成する。
      operationId: create_role
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
        required: true
      responses:
        '201':
          description: ロール作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDetailData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: ロール名重複
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/roles/{role_id}:
    get:
      tags:
      - roles
      summary: GET /api/v1/roles/{role_id}
      description: ロール詳細を取得する。
      operationId: get_role
      parameters:
      - name: role_id
        in: path
        description: ロールID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ロール詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDetailData'
        '404':
          description: ロールが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
    delete:
      tags:
      - roles
      summary: DELETE /api/v1/roles/{role_id}
      description: カスタムロールを削除する。
      operationId: delete_role
      parameters:
      - name: role_id
        in: path
        description: ロールID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: 削除成功
        '404':
          description: ロールが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
    patch:
      tags:
      - roles
      summary: PATCH /api/v1/roles/{role_id}
      description: カスタムロールを更新する。
      operationId: update_role
      parameters:
      - name: role_id
        in: path
        description: ロールID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
        required: true
      responses:
        '200':
          description: ロール更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDetailData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: ロールが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
...
components:
  schemas:
...
    CreateRoleRequest:
      type: object
      description: ロール作成リクエスト
      required:
      - name
      - permissions
      properties:
        name:
          type: string
        description:
          type:
          - string
          - 'null'
        permissions:
          type: array
          items:
            type: string
...
    ProblemDetails:
      type: object
      description: |-
        エラーレスポンス（RFC 9457 Problem Details）

        すべてのサービスで統一されたエラーレスポンス形式。
        `type` フィールドは URI で問題の種類を識別する。
      required:
      - type
      - title
      - status
      - detail
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 0
        detail:
          type: string
...
    RoleDetailData:
      type: object
      description: ロール詳細データ
      required:
      - id
      - name
      - permissions
      - is_system
      - created_at
      - updated_at
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type:
          - string
          - 'null'
        permissions:
          type: array
          items:
            type: string
        is_system:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
    RoleItemData:
      type: object
      description: ロール一覧の要素データ
      required:
      - id
      - name
      - permissions
      - is_system
      - user_count
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type:
          - string
          - 'null'
        permissions:
          type: array
          items:
            type: string
        is_system:
          type: boolean
        user_count:
          type: integer
          format: int64
...
    UpdateRoleRequest:
      type: object
      description: ロール更新リクエスト
      properties:
        name:
          type:
          - string
          - 'null'
        description:
          type:
          - string
          - 'null'
        permissions:
          type:
          - array
          - 'null'
          items:
            type: string
...
  securitySchemes:
    session_auth:
      type: apiKey
      in: cookie
      name: session_id
...